chore: bump to 1.0.0+12; balance-sheet equity + accrual/manufacturing features
- Equity: Owner's Capital + editable Opening Retained Earnings line
  (balance_sheet_entries, equity_recon, BS screen, PDF, l10n, tests)
- Accrued Expenses + Payment Gateway Receivables accounts & dashboards
- COGS analytics dashboard; manufacturing/BOM; books-start cutover
- Firestore rules/indexes for new collections; notification token fix

+++ lib/core/repositories/firestore/firestore_sale_repository.dart
@@ -311,7 +311,7 @@ class FirestoreSaleRepository implements SaleRepository {
           final result = computeStockChange(
             product: product,
             variantId: deduction.variantId,
-            delta: -deduction.quantity,
+            delta: -deduction.quantity.toDouble(),
             valuationMethod: deduction.valuationMethod,
             reason: 'Sale',
           );
@@ -395,7 +395,7 @@ class FirestoreSaleRepository implements SaleRepository {
         final result = computeStockChange(
           product: product,
           variantId: deduction.variantId,
-          delta: -deduction.quantity,
+          delta: -deduction.quantity.toDouble(),
           valuationMethod: deduction.valuationMethod,
           reason: 'Sale',
         );
